PR 625 review Q7: the two fold-taking contributions view a drifted fold at the result's Investable Mask, six fee-clock passages and the example's row-sum rule catch up with #898, ADR 0110 names both clock leaves, and test_06b pins Q6's mask-only type
Piece Q7 of the round-two review of PR 625 (map #746, weight drift and held weights),
reviewed at head `caedfcc6b2` and rebased onto `cf33f7c3ab`. The reference implementation was
run at its moved head and reprints every number `test_11` pins. Six findings, all paid here.

1. `Test.yml` is red at `caedfcc6b2` on `test_06b_asset_panel.jl:645`. Q6's finding 4 made a
   mask-only `feature_matrix` selection stack in the panel's own type and pinned it in
   `test_13b`; `test_06b` (#910) still asserted the old `Float64` fallback on the same shape.
   The assertion reads `Float32`, and its comment states Q6's rule and where the `Float64`
   fallback still lives.

2. `risk_contribution(r, pred::PredictionResult{…, HeldWeightsResult})` on a fold whose result
   carries an Investable Mask raised `DimensionMismatch`: since #856 the record and `res.w`
   come back on the caller's universe, while the result's fee stays on the universe it solved
   on (#892), and #769's method handed the three to the value-level verb unviewed. The fold
   now views the weights and the returns at `result_investable_mask(pred.res)` through
   `investable_weights_view` — whose docstring says a fold's history takes the same column
   view as a path — hands the reduced fee through, and expands the answer with
   `expand_investable_weights`, so a non-investable asset reports exactly `0`, as the
   value-level door does. `plot_risk_contribution(r, pred)` takes the same view.

3. `factor_risk_contribution(r, pred)` paired the fold's reduced `nx` with the expanded record's
   full-width `X` in its default `rd`, and the constructor refused it; the plotting extension's
   factor plot built the same `rd`. `rd` is `Option{<:ReturnsResult} = nothing`, and the new
   per-type `fold_factor_returns(imsk, rd, pred)` reads it by dispatch: a caller's `rd` is on
   the caller's universe and is viewed at the mask through `investable_returns_view`; `nothing`
   builds one from the fold's returns viewed at the same mask, its own `nx` and the factor
   block `reconstruct_rd` carried. The extension's factor plot gains the same keyword and the
   same verb. A new `test_54` testset pins findings 2 and 3 against the same fold solved on the
   hand-reduced universe, and `test_25` draws both plots on a masked fold. The API page gains
   the entry; the manifest row of `01_Base_CrossValidation.jl` moves 67 → 68.

4. Six passages of `17_NetReturnsDrawdowns.jl` still charged the fixed fee every period after
   #898: the plain `calc_net_returns` prose ("one scalar … charged in every period", and an
   illustration whose three net returns all sat `0.0429` below gross — measured now
   `[0.0429, 0.0029, 0.0029]`), the drift docstring's prose and step 5, the per asset twin,
   the path twin, and the `calc_fees` pointer. Each names the two clocks now.

5. Example 08 said a drifted path's rows keep the target's sum. A row sums to
   `1 - (1 - sum(w)) / V[t-1]`, measured `0.7000, 0.7099, 0.7108` on `[0.4, 0.2, 0.1]`; the
   identity that holds is the row plus the deflated cash summing to one, and this page's
   book is fully invested. Both sentences state that.

6. ADR 0110's fee-clock paragraph named "the one leaf `AmortisedFees`" and no scheme field;
   #902 added `FirstObservationFees` and put `fa` on the schemes, recorded in ADR 0122. The
   draft says so in place.

`code_health/size_baseline.toml` records the two rises this commit costs on files held over
the threshold: `01_Base_CrossValidation.jl` 631 → 639 and the plotting extension 2881 → 2888
(#814 stands). Test runs, one process each: `test_11` 297, `test_11b` 252, `test_24` 4442,
`test_33` 135, `test_21a`, `test_09c` green at the head; `test_54` +14, `test_25` +3,
`test_06b`, `test_26`, `test_43`, `test_45`, `test_55`, `test_61` green on this commit;
`size.jl`, `complexity.jl` and `jet.jl` green; example 08 runs as a script and prints its
`#src` numbers.

+++ test/test_06b_asset_panel.jl
@@ -637,12 +637,14 @@ end
     @test eltype(feature_matrix(pcat)) === Float64
     @test eltype(panel_feature_matrix(pcat)[2]) === Float64
 
-    # A selection that names only a mask column derives nothing either.
+    # A selection that names only a mask column derives nothing from the columns it
+    # stacks, so it takes the panel's own type: the fallback is the panel's, and it falls
+    # to Float64 only on a panel with no numeric or tensor field at all (Q6, #802).
     pmsk = asset_panel([NumericPanelInput(; name = "a",
                                           vals = Union{Missing, Float32}[1, missing, 3],
                                           alg = ConstantPanelFill(; val = 0.0f0))])
     @test eltype(feature_matrix(pmsk)) === Float32
-    @test eltype(feature_matrix(pmsk, ["a" => :observed])) === Float64
+    @test eltype(feature_matrix(pmsk, ["a" => :observed])) === Float32
 
     # The Panel Field kinds report the type they contribute.
     @test PortfolioOptimisers.panel_value_eltype(PortfolioOptimisers.panel_field(p32, "a")) ===
